Updates 8-26
- Levels (Enemy Placement, Navmesh, Layout)
- Enemies (Stat Balance)
- Player (Stat Balance, Lazer Update)
- Materials Updated

+++ Boarding Party Redux/Assets/Scripts/SerializedObjects/Gun.cs
@@ -9,7 +9,7 @@ public abstract class Gun : ScriptableObject
     public GameObject gunProjectilePrefab;
     public int gunDamage;
     public int gunProjectileSpeed;
-    public int projectileLifeSpan;
+    public float projectileLifeSpan;
     public float rateOfFire;
     public float nextTimeToFire;
     public Faction hostileFaction;
